Styling Tweaks
Tweaking layout before posting on discord.
Some of the links between pages did not work, so were corrected.
Layout on my portfolio page was not working when the page  resized and the nav bar appeared, fixed by adding a new class and media query .
Added the 'back' button to this page.

Created separate Contact Me page rather than link to the contact section on the index page, although this information is still visible - the same form and styles copied across with little bit of additional text.