Form Validation
Added some elements to my inputs -
icons success / error
Javascript for validation
Styling did not work - took me a while to work this out, but is working.
added a media query and an id to the contact form so when the screen is wider and the side nav bar appears  the form is not hidden behind the nav bar.

tailwind cdn was not added this is why the styling was off